modal is fading background on confirmation

+++ src/app/components/remove-task/remove-task.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { TaskService } from '../../task.service';
 import { Task } from '../../models/task.model';
+import { SimpleModalService } from 'ngx-simple-modal';
+import { ConfirmComponent } from 'src/app/components/confirm/confirm.component';
 
 @Component({
   selector: 'app-remove-task',
@@ -12,9 +14,22 @@ export class RemoveTaskComponent {
   @Input()
   task: Task;
 
-  constructor(private _myService: TaskService) { }
+  constructor(private _myService: TaskService, private simpleModalService: SimpleModalService) { }
 
   removeTask(task: Task) {
-    this._myService.removeTask(task.id);
+    let disposable = this.simpleModalService.addModal(ConfirmComponent, {
+      message: 'Are you sure you want to remove this task?'
+    }).subscribe((isConfirmed) => {
+      if (isConfirmed) {
+        this._myService.removeTask(task.id);
+      }
+    });
+    //We can close modal calling disposable.unsubscribe();
+    //If modal was not closed manually close it by timeout
+    // setTimeout(() => {
+    //   disposable.unsubscribe();
+    // }, 10000);
   }
+
+
 }
